==> 00 - Spanish Wine - Web Scraping.ipynb <==
## Importing the libraries needed
from bs4 import BeautifulSoup
import requests
import pandas as pd
import numpy as np
import re

## Getting the urls
wine_urls = [f'http://elmundovino.elmundo.es/elmundovino/fichavino.html?param={k}' for k in range(0, 25965)]

wine = []
winery = []
country = []
region = []
sub_region = []
year = []
w_type = []
grape = []
alcohol = []
score = []
price = []

for link in wine_urls:
    site = f'{link}'
    resp = requests.get(site)
    soup = BeautifulSoup(resp.text)

    ## Getting the name of the wine
    try:
        name = soup.findAll('font')[10].text.replace('\n','')
        wine.append(name)
    except:
        wine.append('NaN')

    ## Getting the winery name
    try:
        winery_name = soup.findAll('font')[11].text.replace('\n','').replace('Bodega: ','')
        winery.append(winery_name)
    except:
        wine.append('NaN')

    ## Getting the country
    try:
        country_name = soup.findAll('font')[12].text.split('\n')[1]
        country.append(country_name)
    except:
        country.append('NaN')

    ## Getting the region
    try:
        region_name = soup.findAll('font')[12].text.split('\n')[3]
        region.append(region_name)
    except:
        region.append('NaN')

    ## Getting the sub-region
    try:
        sub_region_name = soup.findAll('font')[12].text.split('\n')[4].replace('(','').replace(')','')
        sub_region.append(sub_region_name)
    except:
        sub_region.append('NaN')

    ## Getting the year
    try:
        wine_year = soup.findAll('font')[12].text.split('\n')[6].replace('Año: ','')
        year.append(wine_year)
    except:
        year.append('NaN')

    ## Getting the type
    try:
        wine_type = soup.findAll('font')[12].text.split('\n')[7].replace('Tipo:  ','')
        w_type.append(wine_type)
    except:
        w_type.append('NaN')

    ## Getting the grape(s)
    try:
        grape_type = soup.findAll('font')[12].text.split('\n')[8].replace('Castas: ','')
        grape.append(grape_type)
    except:
        grape.append('NaN')

    ## Getting the alcohol level
    try:
        alcohol_level = soup.findAll('font')[12].text.split('\n')[9].replace('Alcohol: ','').replace('º','')
        alcohol.append(alcohol_level)
    except:
        alcohol.append('NaN')

    ## Getting the review score
    try:
        review_score = soup.findAll('font')[13].text.replace('Puntuación:\n','').replace('/20','')
        score.append(review_score)
    except:
        score.append('NaN')

    ## Getting the price
    try:
        wine_price = soup.findAll('font')[14].text.replace('Precio:\n\n','')[:5]
        price.append(wine_price)
    except:
        price.append('NaN')

## Getting the data into a dataframe
combined = list(zip(wine,winery,country,region,sub_region,year,w_type,grape,alcohol,score,price))
df = pd.DataFrame(combined, columns=['Wine','Winery','Country','Region','Sub Region','Year','Type','Grape','Alcohol','Review Score','Price'])

## Saving the dataframe in a csv file
df.to_csv('winez_0_to_25965.csv', index=False)
