==> 01 - Spanish Wine - Getting the Data.ipynb <==
import pandas as pd

winez = pd.read_csv('winez_0_to_25965.csv')
winez.head()

## Transforming Object variables to Numeric
winez['Price'] = pd.to_numeric(winez['Price'], errors='coerce')
winez['Alcohol'] = pd.to_numeric(winez['Alcohol'], errors='coerce')
winez['Year'] = pd.to_numeric(winez['Year'], errors='coerce')
## Dropping Null values in Price column
winez.dropna(subset='Price',inplace=True)
## Making sure there are no wines at 0€
wines = winez[winez['Price'] >= 1]
## Making sure there are no wines after 2023
wines = wines[wines['Year'] <= 2023]

## Selecting Spanish wines only
spanish = wines[wines['Country']=='España'].copy()
spanish.reset_index(drop=True,inplace=True)

## Dealing with problems during webscraping
spain = spanish[(spanish['Type'] == 'Tinto') | (spanish['Type'] == 'Blanco') | (spanish['Type'] == 'Rosado') | (spanish['Type'] == 'Dulce') | (spanish['Type'] == 'Espumoso') | (spanish['Type'] == 'Generoso')].copy()
## Dropping duplicates
spain.drop_duplicates(inplace=True)
## Removing Country columns as all wines are from Spain now
spain.drop('Country', axis=1, inplace=True)
## Replacing Spanish nomenclature
spain = spain.replace('N/D', 'Unknown')
spain = spain.replace('Tinto','Red').replace('Blanco','White').replace('Rosado','Rose').replace('Espumoso','Sparkling').replace('Dule','Sweet').replace('Generoso','Generous')
spain.head()

## Saving the file
spain.to_csv('spanish_wines.csv', index=False)
